+++ src/locationbox.cc
@@ -1,24 +1,22 @@
 #include "locationbox.h"
 
 #include "events.h"
-#include "inkview.h"
-#include "util.h"
 
-std::array<char, 256> location;
+std::array<char, 256> input_text;
 
 void taranis::keyboard_handler(char *text) {
   const auto event_handler = GetEventHandler();
-  SendEventEx(event_handler, EVT_CUSTOM, CustomEvent::change_location, 0,
-              &location);
+  SendEventEx(event_handler, EVT_CUSTOM, CustomEvent::search_location, 0,
+              &input_text);
 }
 
 void taranis::LocationBox::edit_location() {
   const auto title = GetLangText("Enter location");
-  std::memset(location.data(), '\0', location.size());
+  std::memset(input_text.data(), '\0', input_text.size());
 
-  const auto current_location = format_location(this->model->location);
-  std::memcpy(location.data(), current_location.data(),
-              current_location.size());
-  OpenKeyboard(title, location.data(), location.size() - 1, KBD_NORMAL,
+  const auto formatted_location_name = format_location(this->model->location);
+  std::memcpy(input_text.data(), formatted_location_name.data(),
+              formatted_location_name.size());
+  OpenKeyboard(title, input_text.data(), input_text.size() - 1, KBD_NORMAL,
                keyboard_handler);
 }